# cle-api-swagger.yaml
openapi: 3.0.3
info:
  title: Community Learning Exchange (CLE) API
  description: |
    API specification for the Community Learning Exchange platform that connects learners with teachers.
    
    ## Features
    - User authentication (email/password and Google OAuth)
    - Course management for teachers
    - Availability scheduling
    - Session booking and management
    - Teacher search and discovery
    - Real-time messaging and communication
    
    ## Authentication
    Most endpoints require authentication via JWT token in the Authorization header.
    
  version: 1.0.0
  contact:
    name: CLE Platform
    email: support@cle-platform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.cle-platform.com/v1
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
        teacherId:
          type: string
          format: uuid
        learnerId:
          type: string
          format: uuid
        skillName:
          type: string
        courseId:
          type: string
          format: uuid
        courseName:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [PENDING, CONFIRMED, IN_PROGRESS, COMPLETED, CANCELLED, NO_SHOW]
        format:
          type: string
          enum: [ONLINE, IN_PERSON, HYBRID]
        price:
          type: number
        sessionNotes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        cancellationReason:
          type: string
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid credentials"
        message:
          type: string
          example: "The provided email or password is incorrect"
        code:
          type: string
          example: "AUTH_INVALID_CREDENTIALS"
        timestamp:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
            message: "Authentication token required"
            code: "AUTH_TOKEN_REQUIRED"
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Forbidden"
            message: "Access denied to this resource"
            code: "ACCESS_DENIED"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Not Found"
            message: "The requested resource was not found"
            code: "RESOURCE_NOT_FOUND"
paths:
...
  /sessions:
    get:
      tags:
        - Sessions
      summary: Get user sessions
      description: Get sessions for the authenticated user
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by session status
          schema:
            type: string
            enum: [PENDING, CONFIRMED, IN_PROGRESS, COMPLETED, CANCELLED, NO_SHOW]
        - name: startDate
          in: query
          description: Filter sessions from this date
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Filter sessions until this date
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Sessions
      summary: Book a session
      description: Create a new session booking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                teacherId:
                  type: string
                  format: uuid
                courseId:
                  type: string
                  format: uuid
                availabilityId:
                  type: string
                  format: uuid
                startTime:
                  type: string
                  format: date-time
                endTime:
                  type: string
                  format: date-time
                learnerGoals:
                  type: string
                preparationNotes:
                  type: string
              required:
                - teacherId
                - courseId
                - availabilityId
                - startTime
                - endTime
      responses:
        '201':
          description: Session booked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          description: Invalid booking request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /sessions/{sessionId}:
    get:
      tags:
        - Sessions
      summary: Get session details
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Sessions
      summary: Update session
      description: Update session status or notes
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [PENDING, CONFIRMED, IN_PROGRESS, COMPLETED, CANCELLED, NO_SHOW]
                sessionNotes:
                  type: string
                cancellationReason:
                  type: string
      responses:
        '200':
          description: Session updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Messages
...
security:
  - bearerAuth: []